Changed Member.work-hashmap from <String, Integer> to <String, Double>. Hours worked is not limited to whole hours.

+++ src/TeamOneMiniProject/Member.java
@@ -6,10 +6,10 @@ public class Member {
     private String id;
     private String fullName;
     private int costPerHour;
-    private HashMap<String, Integer> work;
+    private HashMap<String, Double> work;
 
     public Member(){
-        work = new HashMap<String, Integer>();
+        work = new HashMap<String, Double>();
     }
 
     //getters
@@ -25,11 +25,11 @@ public class Member {
         return costPerHour;
     }
 
-    public HashMap<String, Integer> getWork() {
+    public HashMap<String, Double> getWork() {
         return work;
     }
 
-    public Integer getWorkValue(String key){
+    public Double getWorkValue(String key){
         return work.get(key);
     }
